Crate wafel_data_path

Data path compilation and evaluation.

A data path describes a location in memory using a C-like syntax. It allows pointer dereferencing, array indexing, and struct/union field accesses.

There are two types of data paths:

Global: a data path starting from a global variable address

Local: a data path starting from a type, such as a specific struct

It is possible to concatenate paths, either global + local or local + local, if the end type of the first path and the start type of the second path match.

The syntax mostly follows C syntax, e.g. globalVariable.arrayField[3].x for a global path. Local paths are similar, but the global variable is replaced with a type namespace and name, e.g. struct Foo.x, typedef Foo.x.

The crate documentation for wafel_api has more details about the syntax.

Structs

GlobalDataPath

A data path starting from a global variable address.

LocalDataPath

A data path starting from a type, such as a specific struct.

Enums

DataPath

Either a global or a local data path.

DataPathCompileError

DataPathError